Handle HDL with BUILTIN
Lookup of JS chip implementation is only needed for such built-in chips.

In HDLGate's initialize function the wiring up of inputs & outputs for built-in
implementation duplicates much of that for parts.

There's no need for Nand to be a subclass of Gate and have its own inputs &
outputs, because these are now defined by HDL.