Fix bug in creation of internal pins in Definition.Connection
This was erroneously only considering input pins when deciding whether or not
to create a new internal pin. It came to light when I started trying to view
state of the pins when running the test scripts - I saw that spurious internal
pins were created. We'd got away with this because they were never "connected"
to any other pins and so didn't affect the overall behaviour of the top-level
chips.